+++ src/source.extension.cs
@@ -5,7 +5,7 @@ namespace CommentRemover
 		public const string Id = "d7c3f904-cc5a-4d47-aa25-81fb7c36df89";
 		public const string Name = "Comment Remover";
 		public const string Description = "Remove all comments in any file with a click of a button. Can also remove #regions and preserve XML Doc comments.";
-		public const string Version = "1.1";
+		public const string Version = "1.2";
 		public const string Author = "Mads Kristensen";
 		public const string Tags = "comment, cleanup, delete, region, XMLDoc";
 	}